+++ @xen-orchestra/web/src/components/pool/PoolHeader.vue
@@ -30,7 +30,7 @@
     </RouterLink>
     <RouterLink v-slot="{ isActive, href }" :to="`/pool/${pool.id}/vms`" custom>
       <TabItem :active="isActive" :href tag="a">
-        {{ t('vms', 0) }}
+        {{ t('vms', 2) }}
       </TabItem>
     </RouterLink>
   </TabList>